+++ .gitignore
@@ -2,15 +2,6 @@
 .gradle
 /local.properties
 /.idea
-/.idea/caches
-/.idea/libraries
-/.idea/modules.xml
-/.idea/workspace.xml
-/.idea/navEditor.xml
-/.idea/misc.xml
-/.idea/gradle.xml
-/.idea/inspectionProfiles/Project_Default.xml
-/.idea/assetWizardSettings.xml
 .DS_Store
 /build
 /captures
