+++ app/src/main/java/com/example/sprint1homeui/studyRoom/FormatRoom.kt
@@ -1,4 +1,4 @@
-package com.example.sprint1homeui
+package com.example.sprint1homeui.studyRoom
 
 //Imports for the VIEW
